update start game
Former-commit-id: b56a55a67b4c12510c5684d773fa1a5c1f10592f

+++ common/struct_orm.go
@@ -75,7 +75,7 @@ type GameShopItem struct {
 }
 
 type ItemInfo struct {
-	Id             int16 `gorm:"primary_key"`
+	Id             int32 `gorm:"primary_key"`
 	Name           string
 	Type           ItemType
 	Grade          ItemGrade
